Bulletproof

+++ backend/src/queries/users.sql
@@ -2,7 +2,7 @@
 INSERT INTO users (username, display_name, verifier, srp_salt, muk_salt, public_key, private_key, private_key_iv) VALUES (:username, :display_name, :verifier, :srp_salt, :muk_salt, :public_key, :private_key, :private_key_iv) RETURNING id;
 
 /* @name CountByName */
-SELECT COUNT(*) FROM users WHERE username = :username;
+SELECT COUNT(*)::int FROM users WHERE username = :username;
 
 /* @name GetById */
 SELECT id, username FROM users WHERE id = :userId;
